--- utn/pablo_rex/lessons/lesson_12/config/mysql/database.js
@@ -1,13 +1,0 @@
-module.exports = {
-    mysql:{
-        process.env.DB_DATABASE
-        process.env.DB_USER
-        process.env.DB_PASSWORD
-        process.env.DB_HOST
-        process.env.DB_PORT
-        ,
-        mongo:{
-            uri: process.env.MONGO_URI
-        }
-    }
-}
