+++ utn/pablo_rex/lessons/lesson_12/server.js
@@ -11,7 +11,7 @@ app.use(express.json()); //This enable the request.body for POST requests
 //app.use(express.json());    // <==== parse request body as JSON;
 app.use(cookieSession({ name: nameCookie, keys: [keyCookie] }))
 // set the view engine to ejs
-app.set('view engine', 'ejs');
+app.set('view engine', 'hbs');
 
 const apiRoute = require("./routes/api");
 const webRoute = require("./routes/web");
